+++ src/components/ui/slider/index.tsx
